aisentry/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async==0.6.2
asyncio==3.4.3
attrs==23.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
cryptography==42.0.5
dapr==1.13.0
dapr-client==0.4.0b1
deprecation==2.1.0
distro==1.9.0
fastapi==0.110.0
fastapi-responses==0.2.1
Flask==3.0.2
frozenlist==1.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
openai==1.14.2
packaging==24.0
portalocker==2.8.2
priority==2.0.0
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Quart==0.19.4
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
tiktoken==0.5.2
